+++ app/math/vector.ts
@@ -1,0 +1,4 @@
+export class Vector {
+    x: number;
+    y: number;
+}
